feat: move tree diagram to a component. Issue #20
